requirements.txt
python-dotenv
pandas
numpy
mlflow<2.0
certifi
pymongo[srv]
pymongo

-e .